requirements.txt
python-binance
questionary